LOG-3192: refactor visualization into dedicated package(s)
