+++ internal/k8shandler/forwarding.go
@@ -18,6 +18,7 @@ import (
 	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
 	"github.com/openshift/cluster-logging-operator/internal/status"
 	"github.com/openshift/cluster-logging-operator/internal/url"
+	"github.com/openshift/cluster-logging-operator/internal/visualization/console"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
 	client "sigs.k8s.io/controller-runtime/pkg/client"
@@ -355,7 +356,7 @@ func (clusterRequest *ClusterLoggingRequest) LokiStackGatewayService() string {
 // LokiStackURL returns the URL of the LokiStack API for a specific tenant.
 // Returns an empty string if ClusterLogging is not configured for a LokiStack log store.
 func (clusterRequest *ClusterLoggingRequest) LokiStackURL(tenant string) string {
-	service := clusterRequest.LokiStackGatewayService()
+	service := console.LokiStackGatewayService(clusterRequest.Cluster.Spec.LogStore)
 	if service == "" {
 		return ""
 	}
